refactor(access): split principals into store + policy layers (v1.3 review)
Breaks the chat-allowlist ↔ principals circular dependency that the
slice 2 commit papered over with `await import("./principals.js")`.
Both reviewers (Elena architecture, Oliver codegen-failure) flagged it
as a real coupling that would compound under slice 3's capability gate.

Layout:
  principals.ts          — pure I/O on principal records (load/write/
                           list/remove/recordPair). No chat-allowlist
                           dep. capability-bundles is the only import.
  chat-allowlist.ts      — in-memory cache. Imports loadContact from
                           principals.ts (static). Drops the dynamic
                           imports. populate/refresh check
                           `loadContact(c) !== null` directly — at
                           populate time (post-backfill) every cache
                           entry has a principal, so the legacy
                           `isKnownOwner` fallback isn't needed here.
  principals-policy.ts   — NEW. Derived queries combining the two:
                           isContactPermissioned, hasAnyPermissionedContact,
                           getCapabilitiesFor, chatsFor, backfillFromAllowlist.
                           Imports both lower modules.
  index.ts               — barrel re-exports the new module so
                           `access.X` callers see no API change.

The dependency DAG is now: principals → (capability-bundles).
chat-allowlist → principals. principals-policy → both. No cycle.

Tests: 957/0 fail. No public-API change.
